feature/AND/난독화 프로가드 룰 추가(retrofit2, kakao map), 광고용 바텀시트 추가, post item ui text 길이 수정, 북마크 기능 추가, 광고 배너 뷰 추가, 활동지역 없을 경우 분기 추가, 계정탭에서 예약한거 안뜨던 오류수정, edit부분 4번쨰 뷰플립 edittext 포커스 오류 수정, 계정 탭들 무한스크롤 추가, 기타 adapter및 ui수정

+++ app/src/main/java/com/example/mio/ApplyNextActivity.kt
@@ -69,7 +69,7 @@ class ApplyNextActivity : AppCompatActivity() {
         myViewModel = ViewModelProvider(this)[SharedViewModel::class.java]
 
         //val s = savedInstanceState.pu
-        postId = intent.getIntExtra("postId", 0) as Int
+        postId = intent.getIntExtra("postId", 0)
         postData = intent.getSerializableExtra("postData") as PostData
 
         //뒤로가기
